Drop `httpz` for Axum integration
