+++ crates/axum/Cargo.toml
@@ -12,5 +12,8 @@ categories = ["web-programming", "asynchronous"]
 
 [dependencies]
 axum = "0.7.4"
-httpz = { version = "0.0.6", features = ["axum"] }
+cookie = "0.18.0"
+form_urlencoded = "1.2.1"
+futures = "0.3.30"
 rspc = { version = "0.1.4", path = "../..", features = ["internal_axum_07"] }
+serde_json = "1.0.114"
