+++ crates/tauri/Cargo.toml
@@ -12,6 +12,7 @@ categories = ["web-programming", "asynchronous"]
 
 # TODO: Cleanup deps
 [dependencies]
+rspc = { version = "0.1.4", path = "../..", features = [] }
 specta = { version = "1.0.5", features = ["serde", "typescript"] }
 serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.114"
